chore: split init scroll skip from the vm
